+++ src/components/common/RecoilDebugObserver/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './RecoilDebugObserver';
